Rewrite original project (Aug, 2nd 2023)
The game logic, and a basic user interface have been developed into
the project.

Five win conditions have been stablished, and the game ends if any of
this are met OR if a total of 9 moves have been made by the players.
However, the program is still missing the functionality to distinguish
between a player winning and a tie (should be done by next commit).

Gameboard update will also be done by next commit.